move prepare evaluators to transfer matrix
