+++ projects/modular_llm/src/transfer_matrix.py
@@ -1,5 +1,7 @@
+import copy
 import os
 import sys
+from functools import partial
 from typing import Callable, Union
 
 import seaborn as sns
@@ -7,11 +9,17 @@ import wandb
 from matplotlib import pyplot as plt
 from pytorch_lightning import seed_everything
 
+from mttl.datamodule.base import get_datamodule
+
 sys.path.append(os.path.join(os.path.dirname(__file__), "..", "..", "..", ".."))
 
-from mttl.evaluators.evaluators import Evaluator, prepare_evaluator
+from mttl.config import Args, ExpertConfig
+from mttl.evaluators.evaluators import (
+    Evaluator,
+    ExtendedMMLUEvaluator,
+    ExtendedRougeEvaluator,
+)
 from mttl.logging import TableLogger, init_wandb_logger, logger, setup_logging
-from mttl.models.expert_config import ExpertConfig
 
 # register models
 from mttl.models.expert_model import ExpertModel
@@ -19,12 +27,6 @@ from mttl.models.library.expert import Expert
 from mttl.models.library.expert_library import ExpertLibrary
 from mttl.utils import remote_login
 from mttl.vllm_engines.engines import free_memory
-
-DEBUG = False
-if "AMLT_OUTPUT_DIR" in os.environ:
-    DEBUG = False
-if DEBUG:
-    print("!!!!!!!!!!!!!!!!!!!!!! DEBUG MODE")
 
 
 class TransferMatrixConfig(ExpertConfig):
@@ -109,6 +111,53 @@ def eval_all_experts_on_task(
     return log_row
 
 
+def prepare_evaluator(
+    args: Args,
+    dataset,
+    tasks,
+    split=None,
+    subsample=-1,
+    for_generation=None,
+):
+    from mttl.callbacks import TestLossEvaluator
+
+    if args.eval_metric == "loss":
+        EVAL_CLASS = TestLossEvaluator
+        for_generation = for_generation if for_generation is not None else False
+    elif args.eval_metric == "rougeL":
+        EVAL_CLASS = ExtendedRougeEvaluator
+        for_generation = for_generation if for_generation is not None else True
+    elif args.eval_metric == "acc":
+        assert "mmlu" in dataset
+        EVAL_CLASS = ExtendedMMLUEvaluator
+        for_generation = for_generation if for_generation is not None else True
+    else:
+        raise ValueError(f"Unknown eval metric {args.eval_metric}")
+
+    args_copy = copy.deepcopy(args)
+    args_copy.dataset = dataset
+    args_copy.finetune_task_name = tasks
+    args_copy.validation_portion = 0.0
+    dm = get_datamodule(args_copy, for_generation=for_generation)
+
+    if split is not None:
+        evaluator = EVAL_CLASS(
+            datamodule=dm,
+            subsample=subsample,
+            name=tasks,
+            split=split,
+            use_vllm=args.use_vllm,
+        )
+        return evaluator
+
+    return partial(
+        EVAL_CLASS,
+        datamodule=dm,
+        name=tasks,
+        use_vllm=args.use_vllm,
+    )
+
+
 def produce_transfer_matrix(
     args: TransferMatrixConfig,
     expert_lib: ExpertLibrary,
@@ -163,10 +212,6 @@ def run_eval(args: TransferMatrixConfig, debug=None):
     Given a library_id, create transfer matrix: each expert is evaluated on each other expert's dataset.
     """
     seed_everything(args.seed, workers=True)
-    setup_logging(args.output_dir)
-    global DEBUG
-    if debug is not None:
-        DEBUG = debug
 
     if wandb.run is None:
         init_wandb_logger(args)
@@ -187,6 +232,7 @@ def run_eval(args: TransferMatrixConfig, debug=None):
     transfer_table.log_final_table()
     transfer_matrix = transfer_table.df
     transfer_matrix = transfer_matrix.set_index("eval_task")
+
     if wandb.run is not None:
         try:
             _size = 1 * len(transfer_matrix.columns)
@@ -196,6 +242,7 @@ def run_eval(args: TransferMatrixConfig, debug=None):
             wandb.log({"transfer_matrix_heatmap": wandb.Image(ax.get_figure())})
         except Exception as e:
             print(e)
+
     plt.clf()
     print("Transfer matrix", transfer_matrix)
     transfer_matrix.to_csv(os.path.join(args.output_dir, "transfer_matrix.csv"))
